refactor: Create BaseAlgoliaIndex
It contains all functionality common to single-model and
multi-model indexes (aggregations)
